Finishing touches on the Cruso Bow

+++ Content/Foresta/Npcs/Enemies/EnemyPumpkin/EnemyPumpkin.cs
@@ -222,7 +222,7 @@ namespace Crystals.Content.Foresta.Npcs.Enemies.EnemyPumpkin
                 NPC.rotation = MathHelper.Lerp(NPC.rotation, 0, NPC.ai[0]);
                 NPC.velocity +=
                     NPC.DirectionTo(
-                        Vector2.Lerp(attackStartPos, targetPos, (float) EaseFunctions.EaseInBack(NPC.ai[0]))) * 2.5f; 
+                        Vector2.Lerp(attackStartPos, targetPos, EaseFunctions.EaseInBack(NPC.ai[0]))) * 2.5f; 
                 NPC.velocity.X = Math.Clamp(NPC.velocity.X, 0f, 0f);
                 NPC.velocity.Y = Math.Clamp(NPC.velocity.Y, -10f, Single.MaxValue);
             }
@@ -640,7 +640,7 @@ namespace Crystals.Content.Foresta.Npcs.Enemies.EnemyPumpkin
                 NPC.rotation = MathHelper.Lerp(NPC.rotation, 0, NPC.ai[0]);
                 NPC.velocity +=
                     NPC.DirectionTo(
-                        Vector2.Lerp(attackStartPos, targetPos, (float) EaseFunctions.EaseInBack(NPC.ai[0]))) * 2.5f; 
+                        Vector2.Lerp(attackStartPos, targetPos, EaseFunctions.EaseInBack(NPC.ai[0]))) * 2.5f; 
                 NPC.velocity.X = Math.Clamp(NPC.velocity.X, 0f, 0f);
                 NPC.velocity.Y = Math.Clamp(NPC.velocity.Y, -10f, Single.MaxValue);
             }
@@ -990,7 +990,7 @@ namespace Crystals.Content.Foresta.Npcs.Enemies.EnemyPumpkin
                 NPC.rotation = MathHelper.Lerp(NPC.rotation, 0, NPC.ai[0]);
                 NPC.velocity +=
                     NPC.DirectionTo(
-                        Vector2.Lerp(attackStartPos, targetPos, (float) EaseFunctions.EaseInBack(NPC.ai[0]))) * 4f; 
+                        Vector2.Lerp(attackStartPos, targetPos, (EaseFunctions.EaseInBack(NPC.ai[0]))) * 4f); 
                 NPC.velocity.X = Math.Clamp(NPC.velocity.X, 0f, 0f);
                 NPC.velocity.Y = Math.Clamp(NPC.velocity.Y, -10f, 55);
             }
